git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12427 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/gameoptions.cpp
@@ -360,7 +360,7 @@ else {
 	stereo.xSeparation [0] = 65536;
 	stereo.xSeparation [1] = MM2X (RIFT_DEFAULT_IPD);
 	stereo.bChromAbCorr = 1;
-	stereo.nFOV = DEFAULT_FOV;
+	stereo.nFOV = (int) DEFAULT_FOV;
 	}
 }
